Few defination,action , and locators are relocated.
