+++ src/test/java/com/xgen/automation/actions/DeletingConnectionAction.java
@@ -1,11 +1,11 @@
-package com.xgen.automation.actions.CreatingTheConnections;
+package com.xgen.automation.actions;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import org.testng.Assert;
 
 import com.xgen.automation.base.CommonAction;
-import com.xgen.automation.locators.CreatingTheConnections.DeletingtheConnectionsLocator;
+import com.xgen.automation.locators.DeletingtheConnectionsLocator;
 
 public class DeletingConnectionAction extends CommonAction implements DeletingtheConnectionsLocator {
