+++ src/test/java/com/xgen/automation/definations/DeletingTheConnectionDefination.java
@@ -1,8 +1,9 @@
-package com.xgen.automation.definations.CreatingTheConnections;
+package com.xgen.automation.definations;
 
 import io.cucumber.java.en.*;
-import com.xgen.automation.actions.CreatingTheConnections.CreatingConnectionAction;
-import com.xgen.automation.actions.CreatingTheConnections.DeletingConnectionAction;
+
+import com.xgen.automation.actions.CreatingConnectionAction;
+import com.xgen.automation.actions.DeletingConnectionAction;
 import com.xgen.automation.base.CommonAction;
 
 public class DeletingTheConnectionDefination extends CommonAction {
