+++ src/test/java/com/xgen/automation/locators/DeletingtheConnectionsLocator.java
@@ -1,8 +1,8 @@
-package com.xgen.automation.locators.CreatingTheConnections;
+package com.xgen.automation.locators;
 
 import org.openqa.selenium.By;
 
-public interface CreatingtheConnectionsLocator {
+public interface DeletingtheConnectionsLocator {
 	By xx = By.xpath("");
 	By syncIcon= By.xpath("//span[text()='Sync']");
 	By addButtonOnSourcesTab = By.xpath("//p[normalize-space()='Add']");
@@ -11,4 +11,9 @@ public interface CreatingtheConnectionsLocator {
 	
 	By sourceDropDown= By.xpath("//label[text()='Select Source']/following-sibling::div");
 	By destinationDropDown= By.xpath("//label[text()='Select Destination']/following-sibling::div");
+	
+By popupDeleteButton = By.xpath("//p[text()='Delete']");	
+	
+	By popupCancelButton = By.xpath("//p[text()='Cancel']");
+	
 }
